requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0
httpx>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0